rename accordion and locationDetail directories to lowercase
