+++ src/app-components/locationDetail/location-detail.css
